# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LIV LLC accreditation and certificate API
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateInputDesign } from './templateInputDesign';
import type { TemplateInputSettings } from './templateInputSettings';

export interface TemplateInput {
  design: TemplateInputDesign;
  signatory_name: string;
  signatory_title: string;
  /** @nullable */
  validity_months?: number | null;
  settings?: TemplateInputSettings;
}