# lib/api-zod/src/generated/types/traineeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LIV LLC accreditation and certificate API
 * OpenAPI spec version: 1.0.0
 */

export interface TraineeInput {
  first_name: string;
  last_name: string;
  email: string;
  course_name: string;
  completion_date: string;
  grade?: string;
}